代码调整

+++ azz-wx-course-parent/azz-wx-course-pojo/src/main/java/com/azz/wx/course/pojo/vo/ActivityInfo.java
@@ -12,7 +12,7 @@ public class ActivityInfo {
 
     private String activityName;
 
-    private Byte activityStatus;
+    private Integer activityStatus;
 
     private String activityAddress;
